api test, oh it runs fine

+++ modules/models/CRNN_clf.py
@@ -58,7 +58,7 @@ def eval(model:CRNN, dataset:Datasets, params:Params, logger:Logger=None) -> Eva
 
 @torch.inference_mode()
 def infer(model:CRNN, x:Frame, logger:Logger=None) -> Frame:
-  x = torch.from_numpy(x)
+  x = torch.from_numpy(x).float()
   x = x.to(device)
   x = x.unsqueeze(axis=0)   # [B=1, I=96, D=1]
   y = model(x)              # [B=1, NC=4]
